+++ src/js/api/configuration.ts
@@ -1,7 +1,5 @@
-import { restApiPaths } from 'js/restApiPaths';
 import { Region } from 'js/model/Region';
-import { prStore, prAxiosInstance } from "lib/setup";
-import { actions as regionsActions } from "js/redux/regions";
+import { prOnyxiaApiClient } from "lib/setup";
 
 export interface Configuration {
 	regions: Region[];
@@ -14,19 +12,5 @@ export interface Build {
 }
 
 export const getConfiguration = async () => {
-
-	const store = await prStore;
-
-	return (await prAxiosInstance)
-		.get<Configuration>(restApiPaths.configuration)
-		.then(({ data }) => data)
-		.then((resp) => {
-			const configuration = (resp as unknown) as Configuration;
-			store.dispatch(
-				regionsActions.newRegions({
-					"regions": configuration.regions
-				})
-			);
-			return configuration;
-		});
+	return (await prOnyxiaApiClient).getConfigurations()
 };