+++ src/stories/geStory.tsx
@@ -14,12 +14,13 @@ import type { OidcClientConfig, SecretsManagerClientConfig, OnyxiaApiClientConfi
 import type { Props as StoreProviderProps } from "app/interfaceWithLib/StoreProvider";
 import { useTheme } from "@material-ui/core/styles";
 import { RouteProvider } from "app/router";
+import type { Region, Build } from "lib/ports/OnyxiaApiClient";
 
 const { ThemeProvider } = themeProviderFactory(
     { "isReactStrictModeEnabled": false }
 );
 
-const getStoreInitializationParams: StoreProviderProps["getStoreInitializationParams"] = ()=>({
+const getStoreInitializationParams: StoreProviderProps["getStoreInitializationParams"] = () => ({
     "oidcClientConfig": id<OidcClientConfig.Phony>({
         "implementation": "PHONY",
         "tokenValidityDurationMs": Infinity,
@@ -41,7 +42,9 @@ const getStoreInitializationParams: StoreProviderProps["getStoreInitializationPa
     "onyxiaApiClientConfig": id<OnyxiaApiClientConfig.Mock>({
         "implementation": "MOCK",
         "ip": "185.24.1.1",
-        "nomComplet": "John Doe"
+        "nomComplet": "John Doe",
+        regions,
+        build
     })
 });
 
@@ -155,3 +158,75 @@ export function logCallbacks<T extends string>(propertyNames: readonly T[]): Rec
 
 }
 
+
+const regions: Region[] =
+    [
+        {
+            "id": 'datalab',
+            "name": 'DG Insee',
+            "description": 'Region principale. Plateforme hébergée sur les serveurs de la direction générale de l\'INSEE',
+            "location": {
+                "lat": 48.8164,
+                "name": 'Montrouge (France)',
+                "long": 2.3174
+            },
+            "services": {
+                "type": 'MARATHON',
+                "defaultIpProtection": true,
+                "network": 'calico',
+                "namespacePrefix": 'users',
+                "marathonDnsSuffix": 'marathon.containerip.dcos.thisdcos.directory',
+                "expose": {
+                    "domain": 'lab.sspcloud.fr'
+                },
+                "monitoring": {
+                    "URLPattern": 'https://grafana.lab.sspcloud.fr/d/mZUaipcmk/app-generique?orgId=1&refresh=5s&var-id=$appIdSlug'
+                },
+                "cloudshell": {
+                    "catalogId": 'internal',
+                    "packageName": 'shelly'
+                },
+                "initScript": 'https://git.lab.sspcloud.fr/innovation/plateforme-onyxia/services-ressources/-/raw/master/onyxia-init.sh'
+            },
+            "data": {
+                "S3": {
+                    "monitoring": {
+                        "URLPattern": 'https://grafana.lab.sspcloud.fr/d/PhCwEJkMz/minio-user?orgId=1&var-username=$bucketId'
+                    },
+                    "URL": 'https://minio.lab.sspcloud.fr'
+                }
+            }
+        },
+        {
+            "id": 'gke',
+            "name": 'Google cloud',
+            "description": 'Region de test. Aucune garantie de service. A n\'utiliser que pour des tests.',
+            "location": {
+                "lat": 50.8503,
+                "name": 'St. Ghislain (Belgium)',
+                "long": 4.3517
+            },
+            "services": {
+                "type": 'KUBERNETES',
+                "defaultIpProtection": false,
+                "namespacePrefix": 'user-',
+                "expose": {
+                    "domain": 'demo.dev.sspcloud.fr'
+                },
+                "cloudshell": {
+                    "catalogId": 'inseefrlab-helm-charts-datascience',
+                    "packageName": 'cloudshell'
+                }
+            },
+            "data": {
+                "S3": {
+                    "URL": 'https://minio.demo.dev.sspcloud.fr'
+                }
+            }
+        }
+    ];
+
+const build: Build = {
+    "version": "0.7.3",
+    "timestamp": Date.now()
+};